refactor(Drivers): 优化数据类型以节省内存并提高代码效率
- 将 onewire.c 中的 u32 改为 u16，减少内存使用
- 在 time.c 中将 cnt_1s 从 u32 改为 u16，节省内存
- wave.c 和 wave.h 中将 read_d 函数的返回类型从 u32 改为 u16
- 更新 all.h 中的类型定义，移除未使用的 u32 类型

+++ Lanqiao_cup_mcu_template _version4/Drivers/time.c
@@ -35,7 +35,7 @@ void pit1() interrupt 3
 // ne555处理函数
 void ne555_proc()
 {
-    static u32 cnt_1s = 0;
+    static u16 cnt_1s = 0;
     // 每500次中断（即每1s）更新频率计数并清空定时器0
     if (++cnt_1s >= 500) {
         cnt_1s = 0;
